rearrange file structure
