+++ src/components/Book/Book.jsx
@@ -1,4 +1,4 @@
-import "../blocks/Book.css";
+import "./Book.css";
 
 const Book = ({ bookNumber }) => {
   return <div className="book">{`You're on the Book ${bookNumber} page`}</div>;
