+++ src/components/ContactModal/ContactModal.jsx
@@ -1,30 +1,31 @@
 import React from "react";
-import ModalWithForm from "./ModalWithForm";
-import useFormAndValidation from "../hooks/useFormAndValidation";
-import "../blocks/LoginModal.css";
+import ModalWithForm from "../ModalWithForm/ModalWithForm";
+import useFormAndValidation from "../../hooks/useFormAndValidation";
 
-function LoginModal({ titleText, isOpen, isLoading, handleSubmit }) {
+import "./ContactModal.css";
+
+function ContactModal({ titleText, isOpen, isLoading, handleSubmit }) {
   const { values, handleChange, errors, isButtonDisabled } =
     useFormAndValidation(isOpen);
 
   return (
     <ModalWithForm
       titleText={titleText}
-      buttonText={isLoading ? "Logging In..." : "Log In"}
+      buttonText={isLoading ? "Sending..." : "Send"}
       isOpen={isOpen}
       isButtonDisabled={isButtonDisabled}
       handleSubmit={handleSubmit}
     >
       <label
         className={`modal__label ${errors.email ? "modal__error" : ""}`}
-        htmlFor="login-email"
+        htmlFor="contact-email"
       >
-        Email*
+        Your Email*
         <input
           className="modal__input"
           type="email"
           name="email"
-          id="login-email"
+          id="contact-email"
           value={values.email || ""}
           placeholder="Email"
           minLength="1"
@@ -41,30 +42,32 @@ function LoginModal({ titleText, isOpen, isLoading, handleSubmit }) {
         </span>
       </label>
       <label
-        className={`modal__label ${errors.password ? "modal__error" : ""}`}
-        htmlFor="login-password"
+        className={`modal__label ${errors.message ? "modal__error" : ""}`}
+        htmlFor="contact-message"
       >
-        Password*
-        <input
-          className="modal__input"
-          type="password"
-          name="password"
-          id="login-password"
-          value={values.password || ""}
-          placeholder="Password"
+        Your Message*
+        <textarea
+          className="modal__input modal__input_message"
+          type="text"
+          name="message"
+          id="contact-message"
+          value={values.message || ""}
+          placeholder="Message"
+          minLength="1"
+          maxLength="500"
           onChange={handleChange}
           required
         />
         <span
           className={`modal__validation ${
-            errors.password ? "modal__validation_visible" : ""
-          } modal__validation_register-password`}
+            errors.message ? "modal__validation_visible" : ""
+          } modal__validation_register-avatar`}
         >
-          {errors.password}
+          {errors.message}
         </span>
       </label>
     </ModalWithForm>
   );
 }
 
-export default LoginModal;
+export default ContactModal;
