+++ src/components/Header/Header.css
@@ -17,26 +17,3 @@
   margin: 0 auto;
   display: block;
 }
-
-.header__links {
-  display: flex;
-  justify-content: space-between;
-  font-size: 24px;
-  color: white;
-  width: 75%;
-  margin: 20px 0 20px;
-}
-
-.header__link {
-  color: white;
-  text-decoration: none;
-  transition: transform 0.5s;
-  background: none;
-  border: none;
-}
-
-.header__link:hover {
-  transform: scale(1.2);
-  text-decoration: underline;
-  cursor: pointer;
-}
