+++ src/components/LoginModal/LoginModal.jsx
@@ -1,31 +1,30 @@
 import React from "react";
-import ModalWithForm from "./ModalWithForm";
-import useFormAndValidation from "../hooks/useFormAndValidation";
+import ModalWithForm from "../ModalWithForm/ModalWithForm";
+import useFormAndValidation from "../../hooks/useFormAndValidation";
+import "./LoginModal.css";
 
-import "../blocks/ContactModal.css";
-
-function ContactModal({ titleText, isOpen, isLoading, handleSubmit }) {
+function LoginModal({ titleText, isOpen, isLoading, handleSubmit }) {
   const { values, handleChange, errors, isButtonDisabled } =
     useFormAndValidation(isOpen);
 
   return (
     <ModalWithForm
       titleText={titleText}
-      buttonText={isLoading ? "Sending..." : "Send"}
+      buttonText={isLoading ? "Logging In..." : "Log In"}
       isOpen={isOpen}
       isButtonDisabled={isButtonDisabled}
       handleSubmit={handleSubmit}
     >
       <label
         className={`modal__label ${errors.email ? "modal__error" : ""}`}
-        htmlFor="contact-email"
+        htmlFor="login-email"
       >
-        Your Email*
+        Email*
         <input
           className="modal__input"
           type="email"
           name="email"
-          id="contact-email"
+          id="login-email"
           value={values.email || ""}
           placeholder="Email"
           minLength="1"
@@ -42,32 +41,30 @@ function ContactModal({ titleText, isOpen, isLoading, handleSubmit }) {
         </span>
       </label>
       <label
-        className={`modal__label ${errors.message ? "modal__error" : ""}`}
-        htmlFor="contact-message"
+        className={`modal__label ${errors.password ? "modal__error" : ""}`}
+        htmlFor="login-password"
       >
-        Your Message*
-        <textarea
-          className="modal__input modal__input_message"
-          type="text"
-          name="message"
-          id="contact-message"
-          value={values.message || ""}
-          placeholder="Message"
-          minLength="1"
-          maxLength="500"
+        Password*
+        <input
+          className="modal__input"
+          type="password"
+          name="password"
+          id="login-password"
+          value={values.password || ""}
+          placeholder="Password"
           onChange={handleChange}
           required
         />
         <span
           className={`modal__validation ${
-            errors.message ? "modal__validation_visible" : ""
-          } modal__validation_register-avatar`}
+            errors.password ? "modal__validation_visible" : ""
+          } modal__validation_register-password`}
         >
-          {errors.message}
+          {errors.password}
         </span>
       </label>
     </ModalWithForm>
   );
 }
 
-export default ContactModal;
+export default LoginModal;
