+++ src/components/Main/Main.jsx
@@ -1,4 +1,4 @@
-import "../blocks/Main.css";
+import "./Main.css";
 
 const Main = () => {
   return <main className="main"></main>;
