+++ src/components/Preloader/Preloader.jsx
@@ -1,0 +1,7 @@
+import "./Preloader.css";
+
+const Preloader = ({}) => {
+  return <div className="circle-preloader"></div>;
+};
+
+export default Preloader;
