+++ src/components/Profile/Profile.jsx
@@ -1,4 +1,4 @@
-import "../blocks/Profile.css";
+import "./Profile.css";
 
 const Profile = ({}) => {
   return <div className="profile">You're on the Profile page</div>;
