+++ src/components/Store/Store.jsx
@@ -1,4 +1,4 @@
-import "../blocks/Store.css";
+import "./Store.css";
 
 const Store = ({}) => {
   return <div className="store">You're on the Book Store page</div>;
